Removed the InventoryManager class and moved its useful stuff to InventoryManagerMem
