+++ engine/inventory/__init__.py
@@ -1,10 +1,7 @@
-from engine.inventory.inventory import InventoryManager, get_inventory_manager
 from engine.inventory.item import EquippableItem, Item, ItemType
 from engine.inventory.items import ARMORS, FOOD, GROUPTRINKETS, TRINKETS, VALUABLES, WEAPONS
 
 __all__ = [
-    "InventoryManager",
-    "get_inventory_manager",
     "Item",
     "ItemType",
     "EquippableItem",
